Moved nested enum to seperate file
